refactor: restructure app component and remove placeholder content
- Deleted app.component.html and app.component.scss files.
- Updated app.component.ts to use inline template and removed styleUrl.
- Moved app.config.ts to core directory and updated import paths.
- Created empty index files for guards, interceptors, services, features, layouts, shared components, directives, and pipes.
